T15 | QuestionFacoty class and refactor old est
