+++ src/main/java/modelo/options/Option.java
@@ -1,6 +1,5 @@
 package modelo.options;
 
-import com.google.gson.JsonObject;
 import modelo.Points;
 import modelo.scorers.QuestionScorer;
 
@@ -33,13 +32,4 @@ public class Option {
         return this.text;
     }
 
-
-    public static Option unmarshal(JsonObject json){
-        String text = json.get("text").getAsString();
-        boolean isCorrect = json.get("optionScorer").getAsBoolean();
-       if(isCorrect){
-           return new Option(text,new CorrectOptionScorer());
-       }
-        return new Option(text,new IncorrectOptionScorer());
-    }
 }
