+++ src/main/java/modelo/questions/MultipleChoiceQuestion.java
@@ -1,18 +1,12 @@
 package modelo.questions;
 
-import com.google.gson.JsonArray;
-import com.google.gson.JsonElement;
-import com.google.gson.JsonObject;
 import consumables.Consumable;
-import consumables.Multiplicator;
-import consumables.ScoreExclusivity;
-import exceptions.InvalidJsonRecognizerClassException;
 import exceptions.InvalidSizeException;
-import modelo.*;
+import modelo.Player;
+import modelo.Points;
 import modelo.options.Option;
 import modelo.scorers.QuestionScorer;
 
-import java.util.ArrayList;
 import java.util.List;
 
 public class MultipleChoiceQuestion extends Question {
@@ -49,25 +43,4 @@ public class MultipleChoiceQuestion extends Question {
         scorer.score(player,this.points);
     }
 
-    public static Question unmarshal(JsonObject json) throws InvalidJsonRecognizerClassException, InvalidSizeException {
-        try {
-
-            String text = json.get("text").getAsString();
-            String scorerString = json.get("scorer").getAsString();
-
-            List<Option> options = new ArrayList<Option>();
-            JsonArray arrayOptions = json.getAsJsonArray("options");
-            for (JsonElement jsonOption : arrayOptions) {
-                Option option = Option.unmarshal(jsonOption.getAsJsonObject());
-                options.add(option);
-            }
-
-            QuestionScorer questionScorer = selectScorer(scorerString);
-
-            //Question question = question(text, options, questionScorer);
-            return new MultipleChoiceQuestion(text, options, questionScorer, new Multiplicator());
-        } catch (Exception e) {
-            throw e;
-        }
-    }
 }
