+++ src/main/java/modelo/questions/OrderedChoiceQuestion.java
@@ -1,19 +1,11 @@
 package modelo.questions;
 
-import com.google.gson.JsonArray;
-import com.google.gson.JsonElement;
-import com.google.gson.JsonObject;
-import consumables.Multiplicator;
-import exceptions.InvalidJsonRecognizerClassException;
-import exceptions.InvalidSizeException;
 import consumables.Consumable;
 import modelo.Player;
 import modelo.Points;
-import consumables.ScoreExclusivity;
 import modelo.options.Option;
 import modelo.scorers.QuestionScorer;
 
-import java.util.ArrayList;
 import java.util.List;
 
 public class OrderedChoiceQuestion extends Question {
@@ -46,28 +38,6 @@ public class OrderedChoiceQuestion extends Question {
 
         this.consumable.multiplicate(this.points);
         this.scorer.score(player, this.points);
-    }
-
-    public static Question unmarshal(JsonObject json) throws InvalidJsonRecognizerClassException, InvalidSizeException {
-        try {
-
-            String text = json.get("text").getAsString();
-            String scorerString = json.get("scorer").getAsString();
-
-            List<Option> options = new ArrayList<Option>();
-            JsonArray arrayOptions = json.getAsJsonArray("options");
-            for (JsonElement jsonOption : arrayOptions) {
-                Option option = Option.unmarshal(jsonOption.getAsJsonObject());
-                options.add(option);
-            }
-
-            QuestionScorer questionScorer = selectScorer(scorerString);
-
-            //Question question = question(text, options, questionScorer);
-            return new OrderedChoiceQuestion(text, options, questionScorer, new Multiplicator());
-        } catch (Exception e) {
-            throw e;
-        }
     }
 
 }
