+++ src/test/java/modelo/MultipleChoicePenaltyTest.java
@@ -1,15 +1,11 @@
 package modelo;
 
-import com.google.gson.JsonObject;
-import com.google.gson.JsonParser;
 import consumables.Multiplicator;
-import exceptions.InvalidJsonRecognizerClassException;
 import exceptions.InvalidSizeException;
 import modelo.options.CorrectOptionScorer;
 import modelo.options.IncorrectOptionScorer;
 import modelo.options.Option;
 import modelo.questions.MultipleChoiceQuestion;
-import modelo.questions.OrderedChoiceQuestion;
 import modelo.questions.Question;
 import modelo.scorers.PenaltyScorer;
 import modelo.scorers.QuestionScorer;
@@ -165,29 +161,4 @@ public class MultipleChoicePenaltyTest {
         Assert.assertEquals(player.getPoints(), expectedPlayerPoints);
     }
 
-    @Test
-    public void testUnmarshalMultipleChoiceQuestionWithPenaltyAndTestHowItWorks() throws InvalidJsonRecognizerClassException, InvalidSizeException {
-
-        String jString = "{\"text\": \"elegir las opciones que dan como resultado igual a 4\",\"scorer\": \"PenaltyScorer\",\"options\":[{\"text\": \"2 + 2\",\"optionScorer\": true},{\"text\": \"2 * 2\",\"optionScorer\": true},{\"text\": \"1 + 3\",\"optionScorer\": true},{\"text\": \"2^2\",\"optionScorer\": true},{\"text\": \"1 - 3\",\"optionScorer\": false}]}";
-        JsonParser parser = new JsonParser();
-        JsonObject jObj = parser.parse(jString).getAsJsonObject();
-
-        Question question = OrderedChoiceQuestion.unmarshal(jObj);
-
-        Player player = new Player();
-        player.setPoints(7);
-        List<Option> playerOptions = Arrays.asList(
-                new Option("2 + 2", new CorrectOptionScorer()),
-                new Option("5 - 1", new CorrectOptionScorer()),
-                new Option("5 / 1", new IncorrectOptionScorer()));
-        Integer expectedPlayerPoints = 8;
-
-        // When
-        question.selectOptions(playerOptions);
-        question.score(player);
-
-        // Then
-        Assert.assertEquals(player.getPoints(), expectedPlayerPoints);
-
-    }
 }
