+++ src/test/java/modelo/MultipleChoiceTest.java
@@ -1,20 +1,16 @@
 package modelo;
 
-import com.google.gson.JsonObject;
-import com.google.gson.JsonParser;
 import consumables.Multiplicator;
 import consumables.ScoreExclusivity;
-import exceptions.InvalidJsonRecognizerClassException;
 import exceptions.InvalidSizeException;
+import exceptions.NoMoreConsumablesException;
 import modelo.options.CorrectOptionScorer;
 import modelo.options.IncorrectOptionScorer;
 import modelo.options.Option;
 import modelo.questions.MultipleChoiceQuestion;
-import modelo.questions.OrderedChoiceQuestion;
 import modelo.questions.Question;
 import modelo.scorers.MultipleChoiceScorer;
 import modelo.scorers.QuestionScorer;
-import exceptions.NoMoreConsumablesException;
 import org.junit.Assert;
 import org.junit.Rule;
 import org.junit.Test;
@@ -114,33 +110,6 @@ public class MultipleChoiceTest {
         //Act and Assert
         thrown.expect(InvalidSizeException.class);
         Question question = new MultipleChoiceQuestion("Elegir los paises pertenecientes al continente asiatico", options, scorer, new Multiplicator());
-
-    }
-
-
-    @Test
-    public void testUnmarshalMultipleChoiceQuestionAndTestHowItWorks() throws InvalidJsonRecognizerClassException, InvalidSizeException {
-
-        String jString = "{\"text\": \"elegir las opciones que dan como resultado igual a 4\",\"scorer\": \"MultipleChoiceScorer\",\"options\":[{\"text\": \"2 + 2\",\"optionScorer\": true},{\"text\": \"2 * 2\",\"optionScorer\": true},{\"text\": \"1 + 3\",\"optionScorer\": true},{\"text\": \"2^2\",\"optionScorer\": true},{\"text\": \"1 - 3\",\"optionScorer\": false}]}";
-        JsonParser parser = new JsonParser();
-        JsonObject jObj = parser.parse(jString).getAsJsonObject();
-
-        Question question = OrderedChoiceQuestion.unmarshal(jObj);
-
-        Player player = new Player();
-        List<Option> playerOptions = Arrays.asList(
-                new Option("2 + 2", new CorrectOptionScorer()),
-                new Option("2 * 2", new CorrectOptionScorer()),
-                new Option("1 + 3", new CorrectOptionScorer()),
-                new Option("5 - 1", new CorrectOptionScorer()));
-        Integer expectedPlayerPoints = 1;
-
-        // When
-        question.selectOptions(playerOptions);
-        question.score(player);
-
-        // Then
-        Assert.assertEquals(player.getPoints(), expectedPlayerPoints);
 
     }
