+++ src/test/java/modelo/OrderedChoiceTest.java
@@ -1,11 +1,9 @@
 package modelo;
 
-import com.google.gson.JsonObject;
-import com.google.gson.JsonParser;
 import consumables.Multiplicator;
 import consumables.ScoreExclusivity;
-import exceptions.InvalidJsonRecognizerClassException;
 import exceptions.InvalidSizeException;
+import exceptions.NoMoreConsumablesException;
 import modelo.options.CorrectOptionScorer;
 import modelo.options.IncorrectOptionScorer;
 import modelo.options.Option;
@@ -13,7 +11,6 @@ import modelo.questions.OrderedChoiceQuestion;
 import modelo.questions.Question;
 import modelo.scorers.OrderedScorer;
 import modelo.scorers.QuestionScorer;
-import exceptions.NoMoreConsumablesException;
 import org.junit.Assert;
 import org.junit.Test;
 
@@ -74,37 +71,6 @@ public class OrderedChoiceTest {
     }
 
     @Test
-    public void testUnmarshalOrderedChoiceQuestionAndTestHowItWorks() throws InvalidJsonRecognizerClassException, InvalidSizeException {
-
-        String jString = "{\"text\": \"ordene correctamente las siguientes opciones\",\"scorer\": \"OrderedScorer\",\"options\": [{\"text\": \"Primero\",\"optionScorer\": true},{\"text\": \"Segundo\",\"optionScorer\": false},{\"text\": \"Tercero\",\"optionScorer\": false},{\"text\": \"Cuarto\",\"optionScorer\": false}]}";
-        JsonParser parser = new JsonParser();
-        JsonObject jObj = parser.parse(jString).getAsJsonObject();
-
-        Question question = OrderedChoiceQuestion.unmarshal(jObj);
-
-        List<Option> options = Arrays.asList(
-                new Option("Primero", new CorrectOptionScorer()),
-                new Option("Segundo", new IncorrectOptionScorer()),
-                new Option("Tercero", new IncorrectOptionScorer()),
-                new Option("Cuarto", new IncorrectOptionScorer()));
-
-        Player player = new Player();
-        List<Option> playerOptions = Arrays.asList(
-                options.get(0),
-                options.get(1),
-                options.get(3),
-                options.get(2));
-        Integer expectedPlayerPoints = 0;
-
-        // When
-        question.selectOptions(playerOptions);
-        question.score(player);
-
-        // Then
-        Assert.assertEquals(player.getPoints(), expectedPlayerPoints);
-
-    }
-
     public void testGroupChoiceDoublePointsWhenScoreExclusivityActivatedAndOnePlayerAnswerIncorrectly() throws InvalidSizeException, NoMoreConsumablesException {
         // Given
         ScoreExclusivity scoreExclusivity = new ScoreExclusivity();
