+++ src/test/java/modelo/PlayerTest.java
@@ -44,6 +44,15 @@ public class PlayerTest {
         player.activateConsumable(scoreExclusivity);
 
     }
+
+    @Test
+    public void testInstanceAndGetNameShouldBeOK() {
+        String name = "jose";
+        Player player = new Player(name);
+        Assert.assertEquals(name, player.getName());
+
+    }
+
     @Test
     public void testSetAndGetNameShouldBeOK() {
         Player player = new Player();
